# Optimise the log formatter